cleanup: integrate auto evaluation on add button
